@@ -21,7 +21,7 @@
         "source": "https://github.com/yiisoft/error-handler"
     },
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.0",
         "ext-dom": "*",
         "ext-json": "*",
         "alexkart/curl-builder": "^1.0",
@@ -38,6 +38,7 @@
         "httpsoft/http-message": "^1.0.9",
         "phpunit/phpunit": "^9.5",
         "psr/http-server-handler": "^1.0",
+        "rector/rector": "^0.15.10",
         "roave/infection-static-analysis-plugin": "^1.16",
         "spatie/phpunit-watcher": "^1.23",
         "vimeo/psalm": "^4.18",
