@@ -28,12 +28,11 @@ use function array_key_exists;
 use function count;
 use function is_subclass_of;
 use function sprintf;
-use function strpos;
 use function strtolower;
 use function trim;
 
 /**
- * ErrorCatcher catches all throwables from the next middlewares and renders it
+ * `ErrorCatcher` catches all throwables from the next middlewares and renders it
  * according to the content type passed by the client.
  */
 final class ErrorCatcher implements MiddlewareInterface
@@ -46,20 +45,13 @@ final class ErrorCatcher implements MiddlewareInterface
         'text/html' => HtmlRenderer::class,
         '*/*' => HtmlRenderer::class,
     ];
-
-    private ResponseFactoryInterface $responseFactory;
-    private ErrorHandler $errorHandler;
-    private ContainerInterface $container;
     private ?string $contentType = null;
 
     public function __construct(
-        ResponseFactoryInterface $responseFactory,
-        ErrorHandler $errorHandler,
-        ContainerInterface $container
+        private ResponseFactoryInterface $responseFactory,
+        private ErrorHandler $errorHandler,
+        private ContainerInterface $container,
     ) {
-        $this->responseFactory = $responseFactory;
-        $this->errorHandler = $errorHandler;
-        $this->container = $container;
     }
 
     /**
@@ -67,8 +59,6 @@ final class ErrorCatcher implements MiddlewareInterface
      *
      * @param string $contentType The content type to add associated renderers for.
      * @param string $rendererClass The classname implementing the {@see ThrowableRendererInterface}.
-     *
-     * @return self
      */
     public function withRenderer(string $contentType, string $rendererClass): self
     {
@@ -90,8 +80,6 @@ final class ErrorCatcher implements MiddlewareInterface
      *
      * @param string[] $contentTypes The content types to remove associated renderers for.
      * If not specified, all renderers will be removed.
-     *
-     * @return self
      */
     public function withoutRenderers(string ...$contentTypes): self
     {
@@ -113,8 +101,6 @@ final class ErrorCatcher implements MiddlewareInterface
      * Force content type to respond with regardless of request.
      *
      * @param string $contentType The content type to respond with regardless of request.
-     *
-     * @return self
      */
     public function forceContentType(string $contentType): self
     {
@@ -140,11 +126,6 @@ final class ErrorCatcher implements MiddlewareInterface
 
     /**
      * Generates a response with error information.
-     *
-     * @param Throwable $t
-     * @param ServerRequestInterface $request
-     *
-     * @return ResponseInterface
      */
     private function generateErrorResponse(Throwable $t, ServerRequestInterface $request): ResponseInterface
     {
@@ -161,8 +142,6 @@ final class ErrorCatcher implements MiddlewareInterface
      * Returns the renderer by the specified content type, or null if the renderer was not set.
      *
      * @param string $contentType The content type associated with the renderer.
-     *
-     * @return ThrowableRendererInterface|null
      */
     private function getRenderer(string $contentType): ?ThrowableRendererInterface
     {
@@ -176,8 +155,6 @@ final class ErrorCatcher implements MiddlewareInterface
     /**
      * Returns the priority content type from the accept request header.
      *
-     * @param ServerRequestInterface $request
-     *
      * @return string The priority content type.
      */
     private function getContentType(ServerRequestInterface $request): string
@@ -188,7 +165,7 @@ final class ErrorCatcher implements MiddlewareInterface
                     return $header;
                 }
             }
-        } catch (InvalidArgumentException $e) {
+        } catch (InvalidArgumentException) {
             // The Accept header contains an invalid q factor.
         }
 
@@ -204,7 +181,7 @@ final class ErrorCatcher implements MiddlewareInterface
      */
     private function normalizeContentType(string $contentType): string
     {
-        if (strpos($contentType, '/') === false) {
+        if (!str_contains($contentType, '/')) {
             throw new InvalidArgumentException('Invalid content type.');
         }
 
