@@ -9,8 +9,6 @@ use Throwable;
 use Yiisoft\ErrorHandler\ErrorData;
 use Yiisoft\ErrorHandler\ThrowableRendererInterface;
 
-use function get_class;
-
 /**
  * Formats throwable into HTTP headers.
  */
@@ -24,7 +22,7 @@ final class HeaderRenderer implements ThrowableRendererInterface
     public function renderVerbose(Throwable $t, ServerRequestInterface $request = null): ErrorData
     {
         return new ErrorData('', [
-            'X-Error-Type' => get_class($t),
+            'X-Error-Type' => $t::class,
             'X-Error-Message' => $t->getMessage(),
             'X-Error-Code' => (string) $t->getCode(),
             'X-Error-File' => $t->getFile(),
