@@ -19,7 +19,6 @@ use function extract;
 use function file;
 use function file_exists;
 use function func_get_arg;
-use function get_class;
 use function glob;
 use function htmlspecialchars;
 use function implode;
@@ -43,7 +42,6 @@ use function rtrim;
 use function str_replace;
 use function stripos;
 use function strlen;
-use function strpos;
 
 /**
  * Formats throwable into HTML string.
@@ -279,7 +277,7 @@ final class HtmlRenderer implements ThrowableRendererInterface
             }
 
             if (is_object($value)) {
-                $args[$key] = '<span class="title">' . $this->htmlEncode(get_class($value)) . '</span>';
+                $args[$key] = '<span class="title">' . $this->htmlEncode($value::class) . '</span>';
             } elseif (is_bool($value)) {
                 $args[$key] = '<span class="keyword">' . ($value ? 'true' : 'false') . '</span>';
             } elseif (is_string($value)) {
@@ -315,8 +313,6 @@ final class HtmlRenderer implements ThrowableRendererInterface
     /**
      * Renders the information about request.
      *
-     * @param ServerRequestInterface $request
-     *
      * @return string The rendering result.
      */
     public function renderRequest(ServerRequestInterface $request): string
@@ -341,8 +337,6 @@ final class HtmlRenderer implements ThrowableRendererInterface
     /**
      * Renders the information about curl request.
      *
-     * @param ServerRequestInterface $request
-     *
      * @return string The rendering result.
      */
     public function renderCurl(ServerRequestInterface $request): string
@@ -361,8 +355,6 @@ final class HtmlRenderer implements ThrowableRendererInterface
     /**
      * Creates string containing HTML link which refers to the home page
      * of determined web-server software and its full name.
-     *
-     * @param ServerRequestInterface $request
      *
      * @return string The server software information hyperlink.
      */
@@ -397,13 +389,11 @@ final class HtmlRenderer implements ThrowableRendererInterface
     /**
      * Returns the name of the throwable instance.
      *
-     * @param Throwable $throwable
-     *
      * @return string The name of the throwable instance.
      */
     public function getThrowableName(Throwable $throwable): string
     {
-        $name = get_class($throwable);
+        $name = $throwable::class;
 
         if ($throwable instanceof FriendlyExceptionInterface) {
             $name = $throwable->getName() . ' (' . $name . ')';
@@ -439,8 +429,7 @@ final class HtmlRenderer implements ThrowableRendererInterface
 
         $obInitialLevel = ob_get_level();
         ob_start();
-        /** @psalm-suppress InvalidArgument */
-        PHP_VERSION_ID >= 80000 ? ob_implicit_flush(false) : ob_implicit_flush(0);
+        ob_implicit_flush(false);
 
         try {
             $renderer->bindTo($this)($path, $parameters);
@@ -558,7 +547,7 @@ final class HtmlRenderer implements ThrowableRendererInterface
         }
 
         foreach ($this->getVendorPaths() as $vendorPath) {
-            if (strpos($file, $vendorPath) === 0) {
+            if (str_starts_with($file, $vendorPath)) {
                 return true;
             }
         }
@@ -582,7 +571,7 @@ final class HtmlRenderer implements ThrowableRendererInterface
         $rootPath = dirname(__DIR__, 4);
 
         // If the error handler is installed as a vendor package.
-        if (strlen($rootPath) > 6 && strpos($rootPath, 'vendor', -6) !== false) {
+        if (strlen($rootPath) > 6 && str_contains($rootPath, 'vendor')) {
             $this->vendorPaths = [$rootPath];
             return $this->vendorPaths;
         }
