@@ -9,8 +9,6 @@ use Throwable;
 use Yiisoft\ErrorHandler\ErrorData;
 use Yiisoft\ErrorHandler\ThrowableRendererInterface;
 
-use function get_class;
-
 /**
  * Formats throwable into plain text string.
  */
@@ -24,7 +22,7 @@ final class PlainTextRenderer implements ThrowableRendererInterface
     public function renderVerbose(Throwable $t, ServerRequestInterface $request = null): ErrorData
     {
         return new ErrorData(
-            get_class($t) . " with message '{$t->getMessage()}' \n\nin "
+            $t::class . " with message '{$t->getMessage()}' \n\nin "
             . $t->getFile() . ':' . $t->getLine() . "\n\n"
             . "Stack trace:\n" . $t->getTraceAsString()
         );
