@@ -7,15 +7,12 @@ namespace Yiisoft\ErrorHandler\Event;
 use Throwable;
 
 /**
- * ApplicationError represents an application error event.
+ * `ApplicationError` represents an application error event.
  */
 final class ApplicationError
 {
-    private Throwable $throwable;
-
-    public function __construct(Throwable $throwable)
+    public function __construct(private Throwable $throwable)
     {
-        $this->throwable = $throwable;
     }
 
     public function getThrowable(): Throwable
